+++ docs/src/contrib/aiosql/aiosql_contrib.sql
@@ -1,0 +1,5 @@
+-- name: get-username-by-id^
+-- Get username by user's ID.
+SELECT username
+FROM users
+WHERE id = :id;